+++ public/index.php
@@ -5,12 +5,12 @@ use App\Kernel;
 require_once dirname(__DIR__).'/vendor/autoload_runtime.php';
 
 header('Access-Control-Allow-Origin: *');
-header("Access-Control-Allow-Headers: X-API-KEY, Origin, X-Requested-With, Content-Type, Accept, Access-Control-Request-Method");
-header("Access-Control-Allow-Methods: GET, POST, OPTIONS, PUT, PATCH, DELETE");
-header("Allow: GET, POST, OPTIONS, PUT, PATCH, DELETE");
+header('Access-Control-Allow-Headers: X-API-KEY, Origin, X-Requested-With, Content-Type, Accept, Access-Control-Request-Method');
+header('Access-Control-Allow-Methods: GET, POST, OPTIONS, PUT, PATCH, DELETE');
+header('Allow: GET, POST, OPTIONS, PUT, PATCH, DELETE');
 $method = $_SERVER['REQUEST_METHOD'];
-if ($method == "OPTIONS") {
-    die();
+if ('OPTIONS' == $method) {
+    exit;
 }
 
 return function (array $context) {