+++ src/Factory/RedisFactory.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace App\Factory;
+
+use Predis\Client;
+
+class RedisFactory
+{
+    public function __construct(
+        private string $redisAuth
+    ) {
+    }
+
+    public function createRedis(): Client
+    {
+        return new Client($this->redisAuth);
+    }
+}